+++ packages/jest-config/src/Deprecated.ts
@@ -7,7 +7,7 @@
 
 import chalk = require('chalk');
 import type {DeprecatedOptions} from 'jest-validate';
-import prettyFormat from 'pretty-format';
+import {format as prettyFormat} from 'pretty-format';
 
 const format = (value: unknown) => prettyFormat(value, {min: true});