+++ packages/jest-matcher-utils/src/index.ts
@@ -18,7 +18,10 @@ import diffDefault, {
   diffStringsUnified,
 } from 'jest-diff';
 import {getType, isPrimitive} from 'jest-get-type';
-import prettyFormat, {plugins as prettyFormatPlugins} from 'pretty-format';
+import {
+  format as prettyFormat,
+  plugins as prettyFormatPlugins,
+} from 'pretty-format';
 import Replaceable from './Replaceable';
 import deepCyclicCopyReplaceable from './deepCyclicCopyReplaceable';